# swagger.yaml
openapi: 3.0.0
info:
  title: YDX-AI-Backend API
  version: 1.0.0
basePath: /api
components:
  schemas:
...
    IVideos:
      type: object
      properties:
        video_id:
          type: string
        youtube_video_id:
          type: string
        video_name:
          type: string
        video_length:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - video_id
        - youtube_video_id
        - video_name
        - video_length
        - createdAt
        - updatedAt
      additionalProperties: false
...
paths:
...
  /videos/{youtubeId}:
    get:
      summary: Get a video by its YouTube ID
      tags:
        - Videos
      parameters:
        - in: path
          name: youtubeId
          schema:
            type: string
          required: true
          description: The YouTube ID of the video
      responses:
        '200':
          description: A video object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IVideos'
        '404':
          description: The video was not found
  /videos/{youtubeId}/{userId}:
    delete:
      summary: Delete a video for a user by its YouTube ID and user ID
      tags:
        - Videos
      parameters:
        - name: youtubeId
          in: path
          required: true
          schema:
            type: string
          description: The YouTube ID of the video to delete
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the video to delete
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: >-
                      Video with youtubeId: abc123 deleted for user with id:
                      def456
                  message:
                    type: string
                    example: deleted
...